+++ KoenZomers.KeePass.OneDriveSync/OneDriveAboutForm.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Reflection;
 using System.Windows.Forms;
 
 namespace KoenZomersKeePassOneDriveSync
@@ -11,6 +12,10 @@ namespace KoenZomersKeePassOneDriveSync
         public OneDriveAboutForm()
         {
             InitializeComponent();
+
+            var assemblyVersion = Assembly.GetCallingAssembly().GetName().Version;
+
+            HeaderLabel.Text = string.Format("KeePass OneDriveSync v{0}.{1}.{2}", assemblyVersion.Major, assemblyVersion.Minor, assemblyVersion.Build);
         }
 
         private void CloseButton_Click(object sender, EventArgs e)