+++ KoenZomers.KeePass.OneDriveSync/OneDriveConfigForm.Designer.cs
@@ -41,6 +41,7 @@
             this.CloseButton = new System.Windows.Forms.Button();
             this.StatusLabel = new System.Windows.Forms.Label();
             this.AboutButton = new System.Windows.Forms.Button();
+            this.ConfigurationListViewContextItemOpenFileLocation = new System.Windows.Forms.ToolStripMenuItem();
             this.ConfigurationListViewContextMenu.SuspendLayout();
             this.SuspendLayout();
             // 
@@ -104,29 +105,30 @@
             this.ConfigurationListViewContextMenu.Items.AddRange(new System.Windows.Forms.ToolStripItem[] {
             this.ConfigurationListViewContextItemViewDetails,
             this.ConfigurationListViewContextItemSyncNow,
+            this.ConfigurationListViewContextItemOpenFileLocation,
             this.ConfigurationListViewContextItemDelete});
             this.ConfigurationListViewContextMenu.Name = "ConfigurationListViewContextMenu";
-            this.ConfigurationListViewContextMenu.Size = new System.Drawing.Size(138, 70);
+            this.ConfigurationListViewContextMenu.Size = new System.Drawing.Size(169, 114);
             this.ConfigurationListViewContextMenu.Opening += new System.ComponentModel.CancelEventHandler(this.ConfigurationListViewContextMenu_Opening);
             // 
             // ConfigurationListViewContextItemViewDetails
             // 
             this.ConfigurationListViewContextItemViewDetails.Name = "ConfigurationListViewContextItemViewDetails";
-            this.ConfigurationListViewContextItemViewDetails.Size = new System.Drawing.Size(137, 22);
+            this.ConfigurationListViewContextItemViewDetails.Size = new System.Drawing.Size(168, 22);
             this.ConfigurationListViewContextItemViewDetails.Text = "&View Details";
             this.ConfigurationListViewContextItemViewDetails.Click += new System.EventHandler(this.ConfigurationListViewContextItemViewDetails_Click);
             // 
             // ConfigurationListViewContextItemSyncNow
             // 
             this.ConfigurationListViewContextItemSyncNow.Name = "ConfigurationListViewContextItemSyncNow";
-            this.ConfigurationListViewContextItemSyncNow.Size = new System.Drawing.Size(137, 22);
+            this.ConfigurationListViewContextItemSyncNow.Size = new System.Drawing.Size(168, 22);
             this.ConfigurationListViewContextItemSyncNow.Text = "&Sync Now";
             this.ConfigurationListViewContextItemSyncNow.Click += new System.EventHandler(this.ConfigurationListViewContextItemSyncNow_Click);
             // 
             // ConfigurationListViewContextItemDelete
             // 
             this.ConfigurationListViewContextItemDelete.Name = "ConfigurationListViewContextItemDelete";
-            this.ConfigurationListViewContextItemDelete.Size = new System.Drawing.Size(137, 22);
+            this.ConfigurationListViewContextItemDelete.Size = new System.Drawing.Size(168, 22);
             this.ConfigurationListViewContextItemDelete.Text = "&Delete";
             this.ConfigurationListViewContextItemDelete.Click += new System.EventHandler(this.ConfigurationListViewContextItemDelete_Click);
             // 
@@ -164,6 +166,13 @@
             this.AboutButton.Text = "&About";
             this.AboutButton.UseVisualStyleBackColor = true;
             this.AboutButton.Click += new System.EventHandler(this.AboutButton_Click);
+            // 
+            // ConfigurationListViewContextItemOpenFileLocation
+            // 
+            this.ConfigurationListViewContextItemOpenFileLocation.Name = "ConfigurationListViewContextItemOpenFileLocation";
+            this.ConfigurationListViewContextItemOpenFileLocation.Size = new System.Drawing.Size(168, 22);
+            this.ConfigurationListViewContextItemOpenFileLocation.Text = "Open file &location";
+            this.ConfigurationListViewContextItemOpenFileLocation.Click += new System.EventHandler(this.ConfigurationListViewContextItemOpenFileLocation_Click);
             // 
             // OneDriveConfigForm
             // 
@@ -204,5 +213,6 @@
         private System.Windows.Forms.ToolStripMenuItem ConfigurationListViewContextItemSyncNow;
         private System.Windows.Forms.Label StatusLabel;
         private System.Windows.Forms.Button AboutButton;
+        private System.Windows.Forms.ToolStripMenuItem ConfigurationListViewContextItemOpenFileLocation;
     }
 }